+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Update PHP and JS dependencies
 * Add ROLE_PERMANANT to be used for excluding accounts in user cleanup
 * Update ansible roles to fix playbook run
+* Add console command `app:users:list` with option to list inactive users
 
 # 2.7.10 (2021.05.20)